+++ agent/checks/docker.go
@@ -54,6 +54,13 @@ func NewDockerClient(host string, maxbuf int64) (*DockerClient, error) {
 	}, nil
 }
 
+func (c *DockerClient) Close() error {
+	if t, ok := c.client.Transport.(*http.Transport); ok {
+		t.CloseIdleConnections()
+	}
+	return nil
+}
+
 func (c *DockerClient) Host() string {
 	return c.host
 }
@@ -134,7 +141,7 @@ func (c *DockerClient) CreateExec(containerID string, cmd []string) (string, err
 	b, code, err := c.call("POST", uri, data)
 	switch {
 	case err != nil:
-		return "", fmt.Errorf("create exec failed for container %s: %s", containerID, err)
+		return "", fmt.Errorf("create exec failed for container %s: %v", containerID, err)
 	case code == 201:
 		var resp struct{ Id string }
 		if err = json.NewDecoder(bytes.NewReader(b.Bytes())).Decode(&resp); err != nil {
@@ -151,12 +158,18 @@ func (c *DockerClient) CreateExec(containerID string, cmd []string) (string, err
 }
 
 func (c *DockerClient) StartExec(containerID, execID string) (*circbuf.Buffer, error) {
-	data := struct{ Detach, Tty bool }{Detach: false, Tty: true}
+	data := struct{ Detach, Tty bool }{Detach: false, Tty: false}
 	uri := fmt.Sprintf("/exec/%s/start", execID)
 	b, code, err := c.call("POST", uri, data)
 	switch {
-	case err != nil:
-		return nil, fmt.Errorf("start exec failed for container %s: %s", containerID, err)
+	// todo(fs): https://github.com/hashicorp/consul/pull/3621
+	// todo(fs): for some reason the docker agent closes the connection during the
+	// todo(fs): io.Copy call in c.call which causes a "connection reset by peer" error
+	// todo(fs): even though both body and status code have been received. My current is
+	// todo(fs): that the docker agent closes this prematurely but I don't understand why.
+	// todo(fs): the code below ignores this error.
+	case err != nil && !strings.Contains(err.Error(), "connection reset by peer"):
+		return nil, fmt.Errorf("start exec failed for container %s: %v", containerID, err)
 	case code == 200:
 		return b, nil
 	case code == 404:
@@ -164,7 +177,7 @@ func (c *DockerClient) StartExec(containerID, execID string) (*circbuf.Buffer, e
 	case code == 409:
 		return nil, fmt.Errorf("start exec failed since container %s is paused or stopped", containerID)
 	default:
-		return nil, fmt.Errorf("start exec failed for container %s with status %d: %s", containerID, code, b)
+		return nil, fmt.Errorf("start exec failed for container %s with status %d: body: %s err: %v", containerID, code, b, err)
 	}
 }
 
@@ -177,7 +190,7 @@ func (c *DockerClient) InspectExec(containerID, execID string) (int, error) {
 	case code == 200:
 		var resp struct{ ExitCode int }
 		if err := json.NewDecoder(bytes.NewReader(b.Bytes())).Decode(&resp); err != nil {
-			return 0, fmt.Errorf("inspect exec response for container %s cannot be parsed: %s", containerID, err)
+			return 0, fmt.Errorf("inspect exec response for container %s cannot be parsed: %v", containerID, err)
 		}
 		return resp.ExitCode, nil
 	case code == 404: